refactor: split slide and docProps into separate files
