+++ src/slide_voice_app/pptx/pptx_file.py
@@ -1,221 +1,23 @@
 """Central PPTX file model with slide-level operations."""
 
 import tempfile
-import xml.etree.ElementTree as ET
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import Self
 from zipfile import ZIP_DEFLATED, ZipFile
 
-from .audio_model import Audio
-from .audio_read import load_slide_audio
-from .audio_write import delete_slide_audio, upsert_slide_audio
+from .docprops import (
+    count_slides_with_notes,
+    update_app_xml_notes_count,
+    update_core_xml_modified,
+)
 from .exceptions import (
     InvalidPptxError,
-    RelationshipTargetNotFoundError,
     SlideNotFoundError,
 )
-from .namespaces import (
-    NAMESPACE_DCTERMS,
-    NAMESPACE_XSI,
-    REL_TYPE_NOTES_SLIDE,
-    REL_TYPE_SLIDE,
-)
-from .notes import extract_notes_text, write_slide_notes
-from .paths import rels_path_for_path, resolve_target_path, slide_rels_path
-from .rels import (
-    find_relationship_target_by_type,
-    get_relationships_target_by_type,
-    read_rels_path,
-)
-
-
-def _update_core_xml_modified(core_content: bytes) -> bytes:
-    """Update dcterms:modified timestamp in core.xml.
-
-    Args:
-        core_content: Raw bytes of docProps/core.xml.
-
-    Returns:
-        Updated XML bytes.
-    """
-    ET.register_namespace("dcterms", NAMESPACE_DCTERMS)
-    ET.register_namespace("xsi", NAMESPACE_XSI)
-    root = ET.fromstring(core_content)
-    modified = root.find(f"{{{NAMESPACE_DCTERMS}}}modified")
-
-    if modified is None:
-        modified = ET.SubElement(root, f"{{{NAMESPACE_DCTERMS}}}modified")
-
-    now = datetime.now(timezone.utc).replace(microsecond=0)
-    modified.text = now.strftime("%Y-%m-%dT%H:%M:%SZ")
-    modified.set(f"{{{NAMESPACE_XSI}}}type", "dcterms:W3CDTF")
-
-    return ET.tostring(root, encoding="UTF-8", xml_declaration=True)
-
-
-def _update_app_xml_notes_count(app_content: bytes, notes_count: int) -> bytes:
-    """Update Notes count in docProps/app.xml.
-
-    Args:
-        app_content: Raw bytes of docProps/app.xml.
-        notes_count: Number of slides with notes slides.
-
-    Returns:
-        Updated XML bytes.
-    """
-    namespace = (
-        "http://schemas.openxmlformats.org/officeDocument/2006/extended-properties"
-    )
-    root = ET.fromstring(app_content)
-    notes = root.find(f"{{{namespace}}}Notes")
-
-    if notes is None:
-        notes = ET.SubElement(root, f"{{{namespace}}}Notes")
-
-    notes.text = str(notes_count)
-
-    return ET.tostring(root, encoding="UTF-8", xml_declaration=True)
-
-
-def _count_slides_with_notes(work_dir: Path) -> int:
-    """Count slides that reference a notesSlide relationship.
-
-    Args:
-        work_dir: Extracted PPTX workspace root.
-
-    Returns:
-        Number of slides with notes relationships.
-    """
-    rels_dir = work_dir / "ppt/slides/_rels"
-
-    if not rels_dir.exists():
-        return 0
-
-    count = 0
-
-    for rels_path in rels_dir.glob("slide*.xml.rels"):
-        rels_root = ET.fromstring(rels_path.read_bytes())
-
-        if (
-            find_relationship_target_by_type(rels_root, REL_TYPE_NOTES_SLIDE)
-            is not None
-        ):
-            count += 1
-
-    return count
-
-
-class Slide:
-    """Slide model backed by an extracted PPTX workspace."""
-
-    def __init__(
-        self,
-        index: int,
-        slide_path: str,
-        work_dir: Path,
-    ):
-        """Initialize a Slide.
-
-        Args:
-            index: Zero-based slide index.
-            slide_path: OOXML path (e.g. ppt/slides/slide1.xml).
-            work_dir: Extracted PPTX workspace directory.
-        """
-        self.index = index
-        self.slide_path = slide_path
-        self._work_dir = work_dir
-        self._notes = self._read_notes()
-        self.notes_changed = False
-        self.audio: list[Audio] = load_slide_audio(self._work_dir, self.slide_path)
-        self._audio_by_id: dict[str, Audio] = {a.audio_id: a for a in self.audio}
-
-    @property
-    def notes(self) -> str:
-        """Get current in-memory notes text for this slide."""
-        return self._notes
-
-    def set_notes(self, text: str) -> None:
-        """Update in-memory notes text and mark slide as changed.
-
-        Args:
-            text: New plain text notes.
-        """
-        if text != self._notes:
-            self._notes = text
-            self.notes_changed = True
-
-    def save_notes(self) -> None:
-        """Persist notes to workspace if slide notes were edited."""
-        if not self.notes_changed:
-            return
-
-        write_slide_notes(self._work_dir, self.slide_path, self._notes)
-        self.notes_changed = False
-
-    def _read_notes(self) -> str:
-        """Read notes text from extracted workspace files.
-
-        Returns:
-            Notes text as plain string.
-
-        Raises:
-            RelsNotFoundError: If slide relationships are missing.
-            RelationshipTargetNotFoundError: If notes path target cannot be read.
-        """
-        rels_path = slide_rels_path(self._work_dir, self.slide_path)
-        slide_rels = read_rels_path(rels_path)
-
-        if (
-            notes_target := find_relationship_target_by_type(
-                slide_rels,
-                REL_TYPE_NOTES_SLIDE,
-            )
-        ) is None:
-            return ""
-
-        notes_xml_path = resolve_target_path(self.slide_path, notes_target)
-        notes_path = self._work_dir / notes_xml_path
-
-        if not notes_path.exists():
-            raise RelationshipTargetNotFoundError(
-                rels_path_for_path(self.slide_path), notes_xml_path
-            )
-
-        notes_element = ET.fromstring(notes_path.read_bytes())
-        return extract_notes_text(notes_element)
-
-    def _reload_audio(self) -> None:
-        """Reload slide audio from workspace files."""
-        self.audio = load_slide_audio(self._work_dir, self.slide_path)
-        self._audio_by_id = {item.audio_id: item for item in self.audio}
-
-    def add_audio(self, mp3_path: Path) -> None:
-        """Upsert audio for this slide immediately.
-
-        Args:
-            mp3_path: Path to MP3 file.
-
-        Raises:
-            FileNotFoundError: If MP3 file is missing.
-            SlideXmlNotFoundError: If the slide XML file is missing.
-        """
-        upsert_slide_audio(self._work_dir, self.slide_path, mp3_path)
-        self._reload_audio()
-
-    def delete_audio(self, audio_id: str) -> None:
-        """Delete one slide audio entry immediately.
-
-        Args:
-            audio_id: In-memory audio identifier.
-        """
-        audio = self._audio_by_id.get(audio_id)
-
-        if audio is None:
-            return
-
-        delete_slide_audio(self._work_dir, self.slide_path, audio)
-        self._reload_audio()
+from .namespaces import REL_TYPE_SLIDE
+from .paths import resolve_target_path
+from .rels import get_relationships_target_by_type, read_rels_path
+from .slide import Slide
 
 
 class PptxFile:
@@ -384,16 +186,16 @@ class PptxFile:
         if not core_path.exists():
             raise InvalidPptxError(str(self._source_path), "Missing docProps/core.xml")
 
-        core_path.write_bytes(_update_core_xml_modified(core_path.read_bytes()))
+        core_path.write_bytes(update_core_xml_modified(core_path.read_bytes()))
 
         app_path = self._work_dir / "docProps/app.xml"
-        notes_count = _count_slides_with_notes(self._work_dir)
+        notes_count = count_slides_with_notes(self._work_dir)
 
         if not app_path.exists():
             raise InvalidPptxError(str(self._source_path), "Missing docProps/app.xml")
 
         app_path.write_bytes(
-            _update_app_xml_notes_count(app_path.read_bytes(), notes_count)
+            update_app_xml_notes_count(app_path.read_bytes(), notes_count)
         )
 
         with ZipFile(output_path, "w", ZIP_DEFLATED) as zip_file:
